IterativeTargetMethod/requirements.txt
numpy==1.15.3
torchvision==0.2.1
torch==0.4.1
requests==2.20.1
matplotlib==3.0.1
Pillow==5.3.0
